cdk/requirements.txt
annotated-types==0.6.0 
attrs==23.2.0 
aws-cdk-asset-awscli-v1==2.2.202 
aws-cdk-asset-kubectl-v20==2.1.2 
aws-cdk-asset-node-proxy-agent-v6==2.0.1 
aws-cdk-lib==2.105.0 
aws-solutions-constructs-aws-cloudfront-s3==2.46.0 
aws-solutions-constructs-core==2.46.0 
cattrs==23.2.3 
constructs==10.3.0 
importlib-resources==6.1.1 
jsii==1.94.0 
publication==0.0.3 
pydantic-core==2.14.6 
pydantic-settings==2.1.0 
pydantic==2.5.3 
python-dateutil==2.8.2 
python-dotenv==1.0.0 
six==1.16.0 
typeguard==2.13.3 
typing-extensions==4.9.0 
